Move OptionBaseZeroInspectionTests for QuickFixes into relevant files under RD.QuickFixes
