+++ RubberduckTests/QuickFixes/RemoveOptionBaseStatementQuickFixTests.cs
@@ -1,0 +1,101 @@
+﻿using System.Linq;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using RubberduckTests.Mocks;
+using System.Threading;
+using Rubberduck.Inspections.Concrete;
+using Rubberduck.Inspections.QuickFixes;
+using Rubberduck.VBEditor.SafeComWrappers.Abstract;
+using RubberduckTests.Inspections;
+
+namespace RubberduckTests.QuickFixes
+{
+    [TestClass]
+    public class RemoveOptionBaseStatementQuickFixTests
+    {
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void OptionBaseZeroStatement_QuickFixWorks_RemoveStatement()
+        {
+            var inputCode = "Option Base 0";
+            var expectedCode = string.Empty;
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new RedundantOptionInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new RemoveOptionBaseStatementQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void OptionBaseZeroStatement_QuickFixWorks_RemoveStatement_MultipleLines()
+        {
+            var inputCode =
+@"Option _
+Base _
+0";
+
+            var expectedCode = string.Empty;
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new RedundantOptionInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new RemoveOptionBaseStatementQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void OptionBaseZeroStatement_QuickFixWorks_RemoveStatement_InstructionSeparator()
+        {
+            var inputCode = "Option Explicit: Option Base 0: Option Base 1";
+
+            var expectedCode = "Option Explicit: : Option Base 1";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new RedundantOptionInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new RemoveOptionBaseStatementQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void OptionBaseZeroStatement_QuickFixWorks_RemoveStatement_InstructionSeparatorAndMultipleLines()
+        {
+            var inputCode =
+@"Option Explicit: Option _
+Base _
+0: Option Base 1";
+
+            var expectedCode = "Option Explicit: : Option Base 1";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new RedundantOptionInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new RemoveOptionBaseStatementQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+    }
+}
