+++ RubberduckTests/RubberduckTests.csproj
@@ -130,6 +130,7 @@
     <Compile Include="QuickFixes\RemoveEmptyIfBlockQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveExplicitCallStatementQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveExplicitLetStatementQuickFixTests.cs" />
+    <Compile Include="QuickFixes\RemoveOptionBaseStatementQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveTypeHintsQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveUnusedDeclarationQuickFixTests.cs" />
     <Compile Include="QuickFixes\ReplaceEmptyStringLiteralStatementQuickFixTests.cs" />
